third_party/WebKit/LayoutTests/http/tests/devtools/extensions/extensions-reload-expected.txt

CONSOLE MESSAGE: line 1: 42
Tests that webInspector.inspectedWindow.reload() successfully injects and preprocesses user's code upon reload

Started extension.
Running tests...
RUNNING TEST: extension_testReloadInjectsCode
Page reloaded.
Page reloaded.
With injected code: foo = 42
Without injected code: foo = undefined
RUNNING TEST: extension_testReloadInjectsCodeWithMessage
Page reloaded.
Source received:
(function(){console.log(42)})()
All tests done.
